move deployment image to prefect.docker
